feat(user-service-auth): modified the user service to use authorization

+++ Services/User/User.API/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
 using User.API.Dto;
@@ -9,6 +10,7 @@ using User.API.Repositories;
 
 namespace User.API.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("/api/users")]
     public class UserController : Controller
@@ -16,14 +18,17 @@ namespace User.API.Controllers
         private readonly AccountRepository _accountRepository;
         private readonly RoleRepository _roleRepository;
         private readonly IMapper _mapper;
+        private readonly IAuthorizationService _authorizationService;
 
-        public UserController(AccountRepository accountRepository, RoleRepository roleRepository, IMapper mapper)
+        public UserController(AccountRepository accountRepository, RoleRepository roleRepository, IMapper mapper, IAuthorizationService authorizationService)
         {
             _accountRepository = accountRepository;
             _roleRepository = roleRepository;
             _mapper = mapper;
+            _authorizationService = authorizationService;
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Account>>> GetAccounts()
         {
@@ -43,6 +48,10 @@ namespace User.API.Controllers
         {
             var account = await _accountRepository.Get(id);
             if (account == null) return NotFound();
+
+            var result = await _authorizationService.AuthorizeAsync(User, account, "AccountOwner");
+            if (!result.Succeeded) return Forbid();
+            
             account.Role = await _roleRepository.Get(account.RoleId);
             var accountDto = _mapper.Map<GetAccountDto>(account);
             return Ok(accountDto);
@@ -53,6 +62,11 @@ namespace User.API.Controllers
         {
             var account = await _accountRepository.Get(id);
             if (account == null) return NotFound();
+            
+            var result = await _authorizationService.AuthorizeAsync(User, account, "AccountOwner");
+            if (!result.Succeeded) return Forbid();
+            
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             _mapper.Map(updateAccount, account);
             await _accountRepository.Update(account);
             return Ok(_mapper.Map<GetAccountDto>(account));
@@ -63,6 +77,9 @@ namespace User.API.Controllers
         {
             var account = await _accountRepository.Get(id);
             if (account == null) return NotFound();
+            
+            var result = await _authorizationService.AuthorizeAsync(User, account, "AccountOwner");
+            if (!result.Succeeded) return Forbid();
             
             var updateAccount = _mapper.Map<UpdateAccountDto>(account);
             patchDoc.ApplyTo(updateAccount, ModelState);
@@ -75,12 +92,15 @@ namespace User.API.Controllers
             return Ok(_mapper.Map<GetAccountDto>(account));
         }
         
-        
         [HttpDelete("{id:int}")]
         public async Task<ActionResult<Account>> DeleteAccount(int id)
         {
             var account = await _accountRepository.Get(id);
             if (account == null) return NotFound();
+            
+            var result = await _authorizationService.AuthorizeAsync(User, account, "AccountOwner");
+            if (!result.Succeeded) return Forbid();
+            
             await _accountRepository.Delete(account);
             return Ok(_mapper.Map<GetAccountDto>(account));
         }
